requirements.txt
asgiref==3.6.0
Django==4.2
django-admin-sortable==2.3
django-autoslug==1.9.9
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-solo==2.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
Unipath==1.1
